+++ RELEASE_NOTES.md
@@ -1,6 +1,6 @@
 ## Validator Changes
 
-* no changes
+* txTests: honour the tx-ecosystem `lenient-display` test attribute by sending the `lenient-display-validation` parameter with the request
 
 ## Other code changes